requirements.txt
git+https://github.com/huggingface/transformers
accelerate
datasets>=2.16.0,<=3.1.0
accelerate>=0.34.0,<=1.0.1
peft>=0.11.1,<=0.12.0
trl>=0.8.6,<=0.9.6
tokenizers
gradio>=4.38.0,<=5.12.0
pandas>=2.0.0
scipy
einops
sentencepiece
tiktoken
protobuf
uvicorn
pydantic
fastapi
sse-starlette
matplotlib>=3.7.0
fire
packaging
pyyaml
numpy<2.0.0
av
tyro<0.9.0
colorlog
qwen-vl-utils[decord]
torchvision
nltk
rouge-chinese
python-dotenv
bitsandbytes>=0.39.0
evaluate
jiwer
torch>=2.0.0
transformers>=4.35.0
peft>=0.7.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
pillow>=10.0.0
matplotlib>=3.7.0
qwen-vl-utils>=0.0.2
numpy>=1.24.0